requirements.txt
# StocksAgent Dependencies

# Flask for HTTP server
flask>=3.0.0

# LangGraph for agent orchestration
langgraph>=0.2.0

# LangChain core utilities
langchain-core>=0.3.0

# Pydantic for schema validation
pydantic>=2.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Kafka messaging
aiokafka>=0.9.0

# Optional: for real mode Monte Carlo
pathway>=0.8.0          # Streaming data processing
kiteconnect>=4.0.0      # Zerodha API client

# Optional: for future LLM integration
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0

requests==2.32.3

# pin setuptools to less than 81
setuptools<81

pymongo
python-dotenv

spacy
litellm
thefuzz

scikit-learn

openai
joblib